Move normalization function to shared utilities file
We already do other error-handling in here, so this cleans up imports
somewhat.
