+++ x-pack/solutions/security/plugins/security_solution/server/lib/detection_engine/rule_management/api/rules/bulk_actions/bulk_actions_response.ts
@@ -6,33 +6,23 @@
  */
 
 import type { BulkActionSkipResult } from '@kbn/alerting-plugin/common';
-import type { BulkOperationError } from '@kbn/alerting-plugin/server';
 import type { IKibanaResponse, KibanaResponseFactory } from '@kbn/core/server';
-import { transformError } from '@kbn/securitysolution-es-utils';
 import { truncate } from 'lodash';
 import type {
   BulkEditActionResults,
   BulkEditActionSummary,
   NormalizedRuleError,
-  RuleDetailsInError,
 } from '../../../../../../../common/api/detection_engine';
 import type {
   BulkActionType,
   BulkEditActionResponse,
-  BulkActionsDryRunErrCode,
 } from '../../../../../../../common/api/detection_engine/rule_management';
 import { BulkActionTypeEnum } from '../../../../../../../common/api/detection_engine/rule_management';
-import type { PromisePoolError } from '../../../../../../utils/promise_pool';
 import type { RuleAlertType } from '../../../../rule_schema';
-import type { DryRunError } from '../../../logic/bulk_actions/dry_run';
 import { internalRuleToAPIResponse } from '../../../logic/detection_rules_client/converters/internal_rule_to_api_response';
+import { type BulkActionError, normalizeBulkActionError } from './utils';
 
 const MAX_ERROR_MESSAGE_LENGTH = 1000;
-
-export type BulkActionError =
-  | PromisePoolError<string>
-  | PromisePoolError<RuleAlertType>
-  | BulkOperationError;
 
 export const buildBulkResponse = (
   response: KibanaResponseFactory,
@@ -124,7 +114,7 @@ export const normalizeErrorResponse = (errors: BulkActionError[]): NormalizedRul
       message,
       status_code: statusCode,
       err_code: errorCode,
-    } = normalizeError(errorObj);
+    } = normalizeBulkActionError(errorObj);
 
     if (errorsMap.has(message)) {
       errorsMap.get(message)?.rules.push(rule);
@@ -141,42 +131,6 @@ export const normalizeErrorResponse = (errors: BulkActionError[]): NormalizedRul
   return Array.from(errorsMap, ([_, normalizedError]) => normalizedError);
 };
 
-/**
- * Normalize different error types (PromisePoolError<string> | PromisePoolError<RuleAlertType> | BulkOperationError)
- * to one common used in NormalizedRuleError
- * @param error BulkActionError
- */
-const normalizeError = (errorObj: BulkActionError): NormalizedRuleError => {
-  let message: string;
-  let statusCode: number = 500;
-  let errorCode: BulkActionsDryRunErrCode | undefined;
-  let rule: RuleDetailsInError;
-
-  if ('rule' in errorObj) {
-    rule = errorObj.rule;
-    message = errorObj.message;
-  } else {
-    const { error, item } = errorObj;
-    const transformedError =
-      error instanceof Error ? transformError(error) : { message: String(error), statusCode: 500 };
-
-    errorCode = (error as DryRunError)?.errorCode;
-    message = transformedError.message;
-    statusCode = transformedError.statusCode;
-    // The promise pool item is either a rule ID string or a rule object. We have
-    // string IDs when we fail to fetch rules. Rule objects come from other
-    // situations when we found a rule but failed somewhere else.
-    rule = typeof item === 'string' ? { id: item } : { id: item.id, name: item.name };
-  }
-
-  return {
-    message,
-    status_code: statusCode,
-    err_code: errorCode,
-    rules: [rule],
-  };
-};
-
 const getHighestStatusCodeFromErrors = (errors: BulkActionError[]) => {
-  return Math.max(...errors.map((error) => normalizeError(error).status_code));
+  return Math.max(...errors.map((error) => normalizeBulkActionError(error).status_code));
 };
