+++ x-pack/solutions/security/plugins/security_solution/server/lib/detection_engine/rule_management/api/rules/bulk_actions/utils.ts
@@ -7,8 +7,15 @@
 
 import type { ScheduleBackfillResults } from '@kbn/alerting-plugin/server/application/backfill/methods/schedule/types';
 import type { BulkOperationError } from '@kbn/alerting-plugin/server';
+import { transformError } from '@kbn/securitysolution-es-utils';
+import type { RuleDetailsInError } from '../../../../../../../common/api/detection_engine';
+import type {
+  BulkActionsDryRunErrCode,
+  NormalizedRuleError,
+} from '../../../../../../../common/api/detection_engine/rule_management';
 import type { PromisePoolError } from '../../../../../../utils/promise_pool';
 import type { RuleAlertType } from '../../../../rule_schema';
+import type { DryRunError } from '../../../logic/bulk_actions/dry_run';
 
 interface HandleScheduleBackfillResultsParams {
   rules: RuleAlertType[];
@@ -47,4 +54,45 @@ export const handleScheduleBackfillResults = ({
     },
     { backfilled: [], errors } as HandleScheduleBackfillResultsOutcome
   );
+};
+
+export type BulkActionError =
+  | PromisePoolError<string>
+  | PromisePoolError<RuleAlertType>
+  | BulkOperationError;
+
+/**
+ * Normalize different error types (PromisePoolError<string> | PromisePoolError<RuleAlertType> | BulkOperationError)
+ * to one common used in NormalizedRuleError
+ * @param error BulkActionError
+ */
+export const normalizeBulkActionError = (errorObj: BulkActionError): NormalizedRuleError => {
+  let message: string;
+  let statusCode: number = 500;
+  let errorCode: BulkActionsDryRunErrCode | undefined;
+  let rule: RuleDetailsInError;
+
+  if ('rule' in errorObj) {
+    rule = errorObj.rule;
+    message = errorObj.message;
+  } else {
+    const { error, item } = errorObj;
+    const transformedError =
+      error instanceof Error ? transformError(error) : { message: String(error), statusCode: 500 };
+
+    errorCode = (error as DryRunError)?.errorCode;
+    message = transformedError.message;
+    statusCode = transformedError.statusCode;
+    // The promise pool item is either a rule ID string or a rule object. We have
+    // string IDs when we fail to fetch rules. Rule objects come from other
+    // situations when we found a rule but failed somewhere else.
+    rule = typeof item === 'string' ? { id: item } : { id: item.id, name: item.name };
+  }
+
+  return {
+    message,
+    status_code: statusCode,
+    err_code: errorCode,
+    rules: [rule],
+  };
 };
